rename files in tray tool
