+++ openpype/tools/tray/info_widget.py
